Split out appdir; add py34/35 pathlib support
