+++ tomlconf/core.py
@@ -1,67 +1,12 @@
 import os
-import sys
-from pathlib import Path, PosixPath
+from pathlib import Path
+from sys import argv
 
 import tomlkit
 import tomlkit.exceptions as tke
 
+from .appdir import get_app_dir
 from .errors import FileError, TOMLParseError, KeyAlreadyPresent
-
-WIN = sys.platform.startswith('win')
-MAC = sys.platform.startswith('darwin')
-
-
-# get_app_dir is from the Click package. Visit
-# http://click.pocoo.org for more information on the click library.
-
-
-def _posixify(name):
-    """Coverts spaces to dashes; characters to lowercase."""
-    return '-'.join(name.split()).lower()
-
-
-def get_app_dir(app_name, roaming=True, force_posix=False):
-    r"""Returns the config folder for the application.  The default behavior
-    is to return whatever is most appropriate for the operating system.
-    To give you an idea, for an app called ``"Foo Bar"``, something like
-    the following folders could be returned:
-    Mac OS X:
-      ``~/Library/Application Support/Foo Bar``
-    Mac OS X (POSIX):
-      ``~/.foo-bar``
-    Unix:
-      ``~/.config/foo-bar``
-    Unix (POSIX):
-      ``~/.foo-bar``
-    Win XP (roaming):
-      ``C:\Documents and Settings\<user>\Local Settings\Application Data\Foo Bar``
-    Win XP (not roaming):
-      ``C:\Documents and Settings\<user>\Application Data\Foo Bar``
-    Win 7 (roaming):
-      ``C:\Users\<user>\AppData\Roaming\Foo Bar``
-    Win 7 (not roaming):
-      ``C:\Users\<user>\AppData\Local\Foo Bar``
-    app_name (str):
-        the application name.  This should be properly capitalized and
-        can contain whitespace.
-    roaming (bool):
-        controls if the folder should be roaming or not on Windows. Has
-        no affect otherwise.
-    force_posix (bool):
-        if this is set to `True` then on any POSIX system the folder
-        will be stored in the home folder with a leading dot instead of
-        the XDG config home or darwin's application support folder.
-    """
-    if WIN:
-        key = roaming and 'APPDATA' or 'LOCALAPPDATA'
-        return Path(os.environ.get(key, Path.home())) / app_name
-    if force_posix:
-        return PosixPath('~/.' + app_name).expanduser()
-    if MAC:
-        return Path.expanduser('~/Library/Application Support') / app_name
-    return PosixPath(
-        os.environ.get('XDG_CONFIG_HOME', PosixPath('~/.config').expanduser())
-    ) / app_name
 
 
 def parse_toml(s):
@@ -77,40 +22,48 @@ def get_filename(config_path=None, roaming=True, force_posix=False):
     """Return the path/filename where the config will be stored.
 
     When config_path is ...
+
         1. Not Set:
             <appdir>/<progname>/conf.toml
+
         2. App Name (not a directory & doesn't have a file extension):
             <appdir>/<config_path>/conf.toml
+
         3. Path Name: (looks like a directory)
             <config_path>/conf.toml
+
         4. File Name: (has a .toml extension):
             <config_path>
     """
-cpath = Path(config_path) if config_path else None
-kwds = dict(roaming=roaming, force_posix=force_posix)
+    cpath = Path(config_path) if config_path else None
+    kwds = dict(roaming=roaming, force_posix=force_posix)
 
-if not config_path:
-    return get_app_dir(Path(sys.argv[0]).stem, **kwds) / 'conf.toml'
-if cpath.stem == str(cpath):
-    return get_app_dir(cpath.stem, **kwds) / 'conf.toml'
-if not cpath.suffix:
-    return cpath / 'conf.toml'
-if cpath.suffix == '.toml':
-    return cpath
-raise ValueError('Config filename must have a ".toml" extension')
+    if not config_path:
+        return get_app_dir(Path(argv[0]).stem, **kwds) / 'conf.toml'
+    if cpath.stem == str(cpath):
+        return get_app_dir(cpath.stem, **kwds) / 'conf.toml'
+    if not cpath.suffix:
+        return cpath / 'conf.toml'
+    if cpath.suffix == '.toml':
+        return cpath
+    raise ValueError('Config filename must have a ".toml" extension')
 
 
 class Config:
     """File context manager
+
     config_path (str):
         path or file name
+
     mode:
         'r':  read-only (default)
         'r+': read & wite
         'w':  write-only
+
     encoding:
         See the codecs module for the list of supported encodings.
         The default is 'utf-8'.
+
     errors:
         See the documentation for codecs.register for a list of the
         permitted encoding error strings. The default is 'strict'.
@@ -128,7 +81,7 @@ class Config:
         self.__openfile = None
         self._mode = mode
         self.filename = get_filename(config_path, roaming, force_posix)
-        self.path = os.path.split(self.filename)[0]
+        self.path = Path(self.filename).parent
         self.encoding = encoding
         self.errors = errors
         self.data = tomlkit.document()
@@ -139,10 +92,9 @@ class Config:
 
     def __enter__(self):
         try:
-            if not os.path.exists(self.path):
+            if not self.path.exists():
                 os.makedirs(self.path)
-            self.__openfile = open(
-                self.filename,
+            self.__openfile = self.filename.open(
                 mode=self.mode,
                 encoding=self.encoding,
                 errors=self.errors
